requirements.txt
Django==4.2.3
psycopg2-binary==2.9.6
celery==5.1.2
redis==3.5.3
djangorestframework==3.14.0
django-cors-headers==4.3.1